requirements.txt
asgiref==3.5.2
Babel==2.10.1
backports.zoneinfo==0.2.1
blinker==1.4
Django==4.0.4
django-cors-headers==3.11.0
django-createsuperuserwithpassword==2.0.0
django-filter==21.1
django-model-utils==4.2.0
django-money==2.1.1
django-rest-auth==0.9.5
django-seed==0.3.1
djangorestframework==3.13.1
Faker==13.11.1
flake8==4.0.1
gunicorn==20.1.0
importlib-metadata==4.11.3
Markdown==3.3.6
mccabe==0.6.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
py-moneyed==1.2
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.2.4
toposort==1.7
zipp==3.8.0
